requirements.txt
langchain
python-dotenv
langchain_community
pymupdf
langchain-text-splitters
langchain-openai
chromadb
langchain-community
sentence_transformers
langchain_huggingface
faiss-cpu
chromadb
langchain_chroma
gradio
langchain-ollama